+++ CHANGELOG.md
@@ -69,6 +69,7 @@ To learn more about active deprecations, we recommend checking [GitHub Discussio
 
 ### New
 
+- **General**: Introduce new GCP Spanner Scaler ([#XXX](https://github.com/kedacore/keda/issues/XXX))
 - TODO ([#XXX](https://github.com/kedacore/keda/issues/XXX))
 
 #### Experimental